fix platform specific issue
