+++ core/native/src/main/scala/zio/ZIOPlatformSpecific.scala
@@ -20,4 +20,19 @@ import zio.stacktracer.TracingImplicits.disableAutoTrace
 
 private[zio] trait ZIOPlatformSpecific[-R, +E, +A]
 
-private[zio] trait ZIOCompanionPlatformSpecific
+private[zio] trait ZIOCompanionPlatformSpecific {
+
+  /**
+   * Imports a synchronous effect that does blocking IO into a pure value.
+   *
+   * If the returned `ZIO` is interrupted, the blocked thread running the
+   * synchronous effect will be interrupted via `Thread.interrupt`.
+   *
+   * Note that this adds significant overhead. For performance sensitive
+   * applications consider using `attemptBlocking` or
+   * `attemptBlockingCancelable`.
+   */
+  def attemptBlockingInterrupt[A](effect: => A)(implicit trace: ZTraceElement): Task[A] =
+    ZIO.attemptBlocking(effect)
+}
+
