+++ core/shared/src/main/scala/zio/ZIO.scala
@@ -2963,85 +2963,6 @@ object ZIO extends ZIOCompanionPlatformSpecific {
     attemptBlocking(effect).refineToOrDie[IOException]
 
   /**
-   * Imports a synchronous effect that does blocking IO into a pure value.
-   *
-   * If the returned `ZIO` is interrupted, the blocked thread running the
-   * synchronous effect will be interrupted via `Thread.interrupt`.
-   *
-   * Note that this adds significant overhead. For performance sensitive
-   * applications consider using `attemptBlocking` or
-   * `attemptBlockingCancelable`.
-   */
-  def attemptBlockingInterrupt[A](effect: => A)(implicit trace: ZTraceElement): Task[A] =
-    ZIO.suspendSucceed {
-      import java.util.concurrent.atomic.AtomicReference
-      import java.util.concurrent.locks.ReentrantLock
-
-      import zio.internal.OneShot
-
-      val lock   = new ReentrantLock()
-      val thread = new AtomicReference[Option[Thread]](None)
-      val begin  = OneShot.make[Unit]
-      val end    = OneShot.make[Unit]
-
-      def withMutex[B](b: => B): B =
-        try {
-          lock.lock(); b
-        } finally lock.unlock()
-
-      val interruptThread: UIO[Unit] =
-        ZIO.succeed {
-          begin.get()
-
-          var looping = true
-          var n       = 0L
-          val base    = 2L
-          while (looping) {
-            withMutex(thread.get match {
-              case None         => looping = false; ()
-              case Some(thread) => thread.interrupt()
-            })
-
-            if (looping) {
-              n += 1
-              Thread.sleep(math.min(50, base * n))
-            }
-          }
-
-          end.get()
-        }
-
-      blocking(
-        ZIO.uninterruptibleMask(restore =>
-          for {
-            fiber <- ZIO.suspend {
-                       val current = Some(Thread.currentThread)
-
-                       withMutex(thread.set(current))
-
-                       begin.set(())
-
-                       try {
-                         val a = effect
-
-                         ZIO.succeedNow(a)
-                       } catch {
-                         case _: InterruptedException =>
-                           Thread.interrupted // Clear interrupt status
-                           ZIO.interrupt
-                         case t: Throwable =>
-                           ZIO.fail(t)
-                       } finally {
-                         withMutex { thread.set(None); end.set(()) }
-                       }
-                     }.forkDaemon
-            a <- restore(fiber.join).ensuring(interruptThread)
-          } yield a
-        )
-      )
-    }
-
-  /**
    * Locks the specified effect to the blocking thread pool.
    */
   def blocking[R, E, A](zio: => ZIO[R, E, A])(implicit trace: ZTraceElement): ZIO[R, E, A] =
